# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Hospital Management System API.",
        "title": "HMS API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/api/v1",
    "paths": {
...
        "/visits/create": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Doctors can record a visit for a patient.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Visits"
                ],
                "summary": "Record a visit for a patient",
                "parameters": [
                    {
                        "description": "Visit details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.PatientVisitCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/model.PatientVisit"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "$ref": "#/definitions/model.APIError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/model.APIError"
                        }
                    },
                    "404": {
                        "description": "Patient not found",
                        "schema": {
                            "$ref": "#/definitions/model.APIError"
                        }
                    },
                    "500": {
                        "description": "Failed to record patient visit",
                        "schema": {
                            "$ref": "#/definitions/model.APIError"
                        }
                    }
                }
            }
        },
        "/visits/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Doctors and Receptionists can get details of a specific patient visit.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Visits"
                ],
                "summary": "Get details of a specific patient visit",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Visit ID (UUID) for which to get details",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.PatientVisit"
                        }
                    },
                    "400": {
                        "description": "Invalid visit ID format",
                        "schema": {
                            "$ref": "#/definitions/model.APIError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/model.APIError"
                        }
                    },
                    "404": {
                        "description": "Visit not found",
                        "schema": {
                            "$ref": "#/definitions/model.APIError"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Doctors can update patient visit details they recorded. Doctor ID is taken from authenticated user.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Visits"
                ],
                "summary": "Update a specific patient visit",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Visit ID (UUID) for which to update details",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Patient visit details to update",
                        "name": "visit",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.PatientVisitUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.PatientVisit"
                        }
                    },
                    "400": {
                        "description": "Validation failed",
                        "schema": {
                            "$ref": "#/definitions/model.APIError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/model.APIError"
                        }
                    },
                    "404": {
                        "description": "Visit not found",
                        "schema": {
                            "$ref": "#/definitions/model.APIError"
                        }
                    }
                }
            }
        },
        "/visits/{id}/list": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Doctors and Receptionists can list patient visits for a specific patient.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Visits"
                ],
                "summary": "List patient visits for a specific patient",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Patient ID (UUID) for which to list visits",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "description": "Added validation",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "minimum": 0,
                        "type": "integer",
                        "description": "Added validation",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.PaginatedResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid patient ID format",
                        "schema": {
                            "$ref": "#/definitions/model.APIError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/model.APIError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.APIError": {
            "type": "object",
            "properties": {
                "details": {
                    "description": "Can be map[string]string for validation errors, or a simple string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "model.PaginatedResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "The actual list of items"
                },
                "limit": {
                    "description": "The limit used for this page",
                    "type": "integer"
                },
                "offset": {
                    "description": "The offset used for this page",
                    "type": "integer"
                },
                "page": {
                    "description": "Current page number (calculated: offset/limit + 1)",
                    "type": "integer"
                },
                "total": {
                    "description": "Total number of items available across all pages",
                    "type": "integer"
                },
                "total_pages": {
                    "description": "Total pages (calculated: ceil(total/limit))",
                    "type": "integer"
                }
            }
        },
...
        "model.PatientVisit": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "diagnosis": {
                    "type": "string"
                },
                "doctor_id": {
                    "description": "The user ID of the doctor who conducted the visit",
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "notes": {
                    "description": "Additional notes by doctor or about the visit",
                    "type": "string"
                },
                "patient_id": {
                    "type": "string"
                },
                "prescription": {
                    "type": "string"
                },
                "symptoms": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "visit_date": {
                    "description": "Timestamp of the visit",
                    "type": "string"
                }
            }
        },
        "model.PatientVisitCreateRequest": {
            "type": "object",
            "required": [
                "patient_id"
            ],
            "properties": {
                "diagnosis": {
                    "type": "string"
                },
                "notes": {
                    "type": "string"
                },
                "patient_id": {
                    "type": "string"
                },
                "prescription": {
                    "type": "string"
                },
                "symptoms": {
                    "type": "string"
                },
                "visit_date": {
                    "description": "Expect RFC3339. If empty, handler might default to time.Now()",
                    "type": "string"
                }
            }
        },
        "model.PatientVisitUpdateRequest": {
            "type": "object",
            "properties": {
                "diagnosis": {
                    "type": "string"
                },
                "notes": {
                    "type": "string"
                },
                "prescription": {
                    "type": "string"
                },
                "symptoms": {
                    "type": "string"
                },
                "visit_date": {
                    "type": "string"
                }
            }
        },
...
    }
}